Small refactorization csv to JSON; divide component Administrator
