+++ src/components/utils/backend-links.ts
@@ -4,6 +4,6 @@ export const URL_FORGOT_PASSWORD = 'http://localhost:3001/user/change-password';
 
 export const URL_CHANGE_PASSWORD = 'http://localhost:3001/user/new-password';
 
+export const URL_IMPORT_USERS ='http://localhost:3001/user/import';
 
-export const URL_IMPORT_USERS ='http://localhost:3001/'
-export const URL_ADD_HR = 'http://localhost:3001/user/add-hr'
+export const URL_ADD_HR = 'http://localhost:3001/user/add-hr';
